+++ internal/origin_leaf/leaf/spaces.go
@@ -6,11 +6,14 @@ import (
 )
 
 // update the space on a leaf node
-func (s *Session) UpdateSpace(space *model.Space) bool {
+func (s *Session) UpdateSpace(space *model.Space, updateFields []string) bool {
 	if s.token == nil || space.UserId == s.token.UserId {
 		message := &msg.ClientMessage{
 			Command: msg.MSG_UPDATE_SPACE,
-			Payload: space,
+			Payload: &msg.UpdateSpace{
+				Space:        *space,
+				UpdateFields: updateFields,
+			},
 		}
 
 		s.ch <- message
@@ -32,23 +35,27 @@ func (s *Session) DeleteSpace(id string) {
 }
 
 // update the space on all leaf nodes
-func UpdateSpace(space *model.Space) {
+func UpdateSpace(space *model.Space, updateFields []string, skipSession *Session) {
 	sessionMutex.RLock()
 	defer sessionMutex.RUnlock()
 
 	// Send the space to all followers
 	for _, session := range session {
-		session.UpdateSpace(space)
+		if session != skipSession {
+			session.UpdateSpace(space, updateFields)
+		}
 	}
 }
 
 // delete the space on all leaf nodes
-func DeleteSpace(id string) {
+func DeleteSpace(id string, skipSession *Session) {
 	sessionMutex.RLock()
 	defer sessionMutex.RUnlock()
 
 	// Send the user to all followers
 	for _, session := range session {
-		session.DeleteSpace(id)
+		if session != skipSession {
+			session.DeleteSpace(id)
+		}
 	}
 }
